requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.5
django-4-jet==1.0.9
django-appconf==1.0.5
django-compressor==4.4
django-nested-admin==4.0.2
django-solo==2.1.0
django-tinymce==3.6.1
gunicorn==21.2.0
idna==3.4
mysqlclient==2.2.4
packaging==23.1
Pillow==10.0.1
pycryptodome==3.20.0
python-dotenv==1.0.0
python-monkey-business==1.0.0
razorpay==1.4.2
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
stripe==7.0.0
typing_extensions==4.8.0
urllib3==2.0.7